backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
PyJWT[crypto]==2.11.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.22
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
